backend/requirements.txt
# Core ML libraries
tensorflow>=2.10.0
opencv-python>=4.5.0
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
rarfile>=4.0
pathlib

# Web framework
Flask>=2.0.0
Flask-CORS>=3.0.0

# Image processing
Pillow>=8.0.0

# Utilities
python-dotenv>=0.19.0
gunicorn>=20.0.0